backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
pymongo==4.5.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
pyjwt==2.8.0
dnspython==2.4.2
pytest==7.4.0
openai==1.3.0
beautifulsoup4==4.12.2


